requirements.txt
PySwitchbot>=2.0.0
